feat: interfaces-b11-cli-work-production-fan-in-005 - Verify contract/package gates and Batch 12 non-goals
